Função Venda e Movimentação de estoque

+++ src/main/java/com/suicidaesquadrao/controle/produtoControle.java
@@ -4,10 +4,7 @@ package com.suicidaesquadrao.controle;
 import com.suicidaesquadrao.dao.produtoDAO;
 import com.suicidaesquadrao.model.produtos;
 import java.io.IOException;
-import java.io.PrintWriter;
 import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.List;
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
@@ -15,10 +12,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import util.validacaoException;
 
-/**
- *
- * @author Deise
- */
+
 public class produtoControle extends HttpServlet {
 
     private produtoDAO ProdutoDAO = new produtoDAO();
@@ -84,7 +78,7 @@ public class produtoControle extends HttpServlet {
         
         //preencher campos
         String nome= request.getParameter("nome");
-        float quantidade=Float.parseFloat(request.getParameter("quantidade"));
+        int quantidade=Integer.parseInt(request.getParameter("quantidade"));
         double preco=Double.parseDouble(request.getParameter("preco"));
         int filial=Integer.parseInt(request.getParameter("filial"));
         String id = request.getParameter("id");
